pyproject.toml
[project]
name = "langchain-dev-utils"
version = "1.0.0rc1"
description = "A practical utility library for LangChain and LangGraph development"
readme = "README.md"
authors = [{ name = "tiebingice", email = "tiebingice123@outlook.com" }]
requires-python = ">=3.11"
dependencies = ["langchain>=1.0.0rc2", "langgraph>=1.0.0rc1"]

[project.urls]
"Source Code" = "https://github.com/TBice123123/langchain-dev-utils"
repository = "https://github.com/TBice123123/langchain-dev-utils"
documentation = "https://tbice123123.github.io/langchain-dev-utils-docs/en/"

[project.optional-dependencies]
standard = ["langchain-openai>=1.0.0a1"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[dependency-groups]
tests = [
    "langchain-community>=1.0.0a1",
    "python-dotenv>=1.1.1",
    "langchain-tests>=1.0.0a2",
]
